[+] moved AqaGroupDynamicButtonsTest -> AqaGroupEviltesterTest
