+++ src/test/java/school/redrover/AqaGroupEviltesterTest.java
@@ -6,11 +6,13 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.openqa.selenium.support.ui.WebDriverWait;
 import org.testng.Assert;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Ignore;
 import org.testng.annotations.Test;
 
+import java.time.Duration;
 import java.util.Map;
 
 public class AqaGroupEviltesterTest extends AqaGroupBaseTest {
@@ -18,6 +20,31 @@ public class AqaGroupEviltesterTest extends AqaGroupBaseTest {
     private static final String BUTTONS_URL = "https://testpages.eviltester.com/styled/dynamic-buttons-disabled.html";
     private static final String ALERTS_URL = "https://testpages.eviltester.com/styled/alerts/alert-test.html";
     private static final String FAKE_ALERTS_URL = "https://testpages.eviltester.com/styled/alerts/fake-alert-test.html";
+
+    @Test
+    public void testSimpleDynamicButtons() {
+        getDriver().get("https://testpages.eviltester.com/styled/dynamic-buttons-simple.html");
+
+        getDriver().findElement(By.id("button00")).click();
+        getDriver().findElement(By.id("button01")).click();
+        WebDriverWait wait = new WebDriverWait(getDriver(), Duration.ofSeconds(5));
+        wait.until(ExpectedConditions.presenceOfElementLocated(By.id("button02")));
+        getDriver().findElement(By.id("button02")).click();
+        wait.until(ExpectedConditions.presenceOfElementLocated(By.id("button03")));
+        getDriver().findElement(By.id("button03")).click();
+
+        Assert.assertEquals(getDriver().findElement(By.id("buttonmessage")).getText(), "All Buttons Clicked", "Not All Buttons Clicked");
+    }
+
+    @Test
+    public void testXHTTPMessages() {
+        getDriver().get("https://testpages.eviltester.com/styled/sync/xhttp-messages.html");
+
+        WebDriverWait wait = new WebDriverWait(getDriver(), Duration.ofSeconds(60));
+        wait.until(ExpectedConditions.textToBePresentInElement(getDriver().findElement(By.id("messagesstatus")), "Message Count: 0 : 0"));
+
+        Assert.assertEquals(getDriver().findElement(By.id("messagesstatus")).getText(), "Message Count: 0 : 0", "message count is not zero");
+    }
 
     @Test
     public void testDisabledDynamicButtons() {
